Flowc nim dynamic (#1166)
* flowc: nim: type index / type table are used

* flowc: nim: makeStructValue works

* flowc: nim: almost all functions from 'dynamic' are implemented

* flowc: duplicate imports are not allowed

* flowc: nim: endless recursion in startsWith is fixed

* flowc: nim: debug prints are disabled

* flowc: fi_helpers2 is added

* flowc: nim: some enhancements

* flowc: nim: case of no structs is considered

+++ tools/flowc/backends/nim/natives/ds/array/iteriUntil.nim
@@ -2,6 +2,6 @@
 # Returns index of last element function was applied to.
 proc iteriUntil*[T](a: openArray[T], op: proc(idx: int32, v: T): bool): int32 =
   for i in 0..a.len - 1:
-    if op(i, a[i]):
-      return i
-  return a.len
+    if op(int32(i), a[int32(i)]):
+      return int32(i)
+  return int32(a.len)
